Workaround for race condition between yarn install and component-library build (#168)
There is a known Yarn issue in which the installation of dependencies utilizing prepare scripts results in multiple yarn processes running concurrently and one of them inevitably failing. We still need to build component-library on install though (assuming we want to iteratively and quickly develop against it at times), so we're now getting around this by throttling yarn's concurrent request cap.

Most definitely a workaround, but it works for now.