Auto-generate users present in WingedKeys but missing from Fawkes (#334)
I know this solution isn't necessarily good, but I've been trying to figure out if it introduces any vulnerabilities and I honestly can't think of any. That being said, we should def do away with this once we have an actual user management solution in place. 😅

If somebody makes a request with a JWT that, when decoded, refers to a user that Fawkes isn't familiar with, we ask Winged Keys for more details about the user. If there is in fact a Winged Keys user corresponding to the decoded JWT's sub that isn't present in Fawkes, just add them to Fawkes.

If there's a huge vulnerability that I'm overlooking here, PLEASE let me know and I'll blow this all away. 🙃