Add permission gate to sites and funding spaces when changing child enrollments (#464)
This PR ensures that when someone attempts to change a child's enrollment, the sites and funding spaces they have available are only pulled from the child's associated organization. The sites and funding spaces made available during initial child submission into the system are already limited to the organizations the user has access to.