Enable e2e encryption on Elastic Beanstalk (#723)
The Application Load Balancer in front of all of our EC2s is now responsible for redirecting HTTP traffic to HTTPS, so we no longer need to have NGINX worry about that. We DO unfortunately need to manually remove the default security group that Elastic Beanstalk always creates for it's EC2s, because (funnily enough) it's less secure than what our requirements necessitate.