Fix seemingly transient enrollment failure issue (#889)
When the changeEnrollment route is invoked, the site ID supplied with the new enrollment was previously provided as a string (even though the actual Site id DB column is an auto-incrementing int). This normally wouldn't be all that much of an issue but, for whatever reason, whenever that site ID exceeds 3 digits, our API interprets it as a Moment instance instead of a number. So, for example, an enrollment site ID of 2077 supplied in the request body of the client is received by the changeEnrollment API as a Moment instance of January 1st, 2077. Of course this breaks immediately when trying to use site ID in the where clause of a findOne query.

The quickest fix for now is to just have the client specify enrollment.site.id as a number, not a string.