First pass at preview screen (#1192)
As part of the change from the Upload page to an Upload Wizard, we need a screen which shows users a preview of their upload, including the type of upload/update performed as well as whether any records are missing info. This PR creates the Preview screen, which includes the roster preview table, altered header copy, and cards for number of uploaded records of each type.