Remove nested immer vulnerability in react-scripts dependency (#1193)
Updates our version of react-scripts to address a known vulnerability with a nested dependency of that package (similar issue captured here, too). Necessary so that the audit portion of our CircleCI builds succeed again (the front-end yarn audit step was previously failing 100% of the time), and we can continue running our unit tests regularly.

Also updates snapshots/unit tests that were previously failing but we didn't know about because of the audit failures.