Make day and time equivalent in funding type mappings (#1234)
Our users interchangeably use 'day' and 'time' to describe funding types when speaking of full-day, part-time, etc. However, the app treats this as different things, and so flags pretty dumb things as errors when users upload their sheets. This PR forces the various permutations of full/part and day/time to map equivalently to the same root so they are safely processed no matter which form the user chooses.