Make RP months align with user entered dates (#1232)
Reporting periods suck because they don't align with normal month by month dates, and this makes a big headache for users. The mapping between reporting period dates and the dates they write in their sheets is completely opaque and does not need to cause them stress. This PR address a weakness in reporting period mapping by:

1.  not automatically mapping all parsed reporting period fields to have 'startOfMonth' as their day value
2.  first checking if a user supplied date is "meaningful" and finding a reporting period whose start and end dates sandwich the user date (makes more sense in the code comments)
3.  if a user didn't supply a day value, making sure that the found reporting period is the right one to cover their given reporting period date