Fix funding space seeder bug (#1352)
This PR fixes a race condition with our test environment seeder. Previously, we decided whether or not to create funding spaces for a particular organization based on if there were any rows in the funding space table at all. This was problematic, because occasionally the creation of funding spaces for one organization would take place prior to the creation of funding spaces for another. And in those scenarios, because we already had data in that table, we would simply not create any funding spaces for all remaining organizations.

We no longer do that, instead checking to see if there are any funding spaces for the organization we're currently dealing with to determine whether or not they need to be created.

ALSO, I shimmied in a tiny fix to the rebuild.sh script, from when it was since moved to the scripts directory.